LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/get_all_cookies-default-samesite-expected.txt


FAIL Get all cookies w/ default SameSite assert_equals: expected "Lax" but got "None"
